upgrade containerd to 2.0 for al2023 over 1.33

+++ templates/al2023/provisioners/cache-pause-container.sh
@@ -4,6 +4,10 @@ set -o nounset
 set -o errexit
 set -o pipefail
 
+if vercmp "$KUBERNETES_VERSION" gteq "1.33.0"; then
+    exit 0
+fi
+
 sudo systemctl start containerd
 cache-pause-container -i ${PAUSE_CONTAINER_IMAGE}
 sudo systemctl stop containerd
